+++ src/functionalTest/kotlin/com/autonomousapps/jvm/projects/SecurityProviderAdvice.kt
@@ -1,0 +1,7 @@
+package com.autonomousapps.jvm.projects
+
+import com.autonomousapps.advice.Advice
+
+object SecurityProviderAdvice {
+  val expectedAdvice = listOf<Advice>()
+}